+++ sigil/__init__.py
@@ -1,6 +1,3 @@
-import os
+from importlib.metadata import version
 
-__version__ = '0.0.0'
-dirname = os.path.dirname(__file__)
-with open(os.path.join(dirname, '..', 'version.txt')) as f:
-    __version__ = f.read().strip()
+__version__ = version("sigil")