+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+* Change dependency manager to Poetry
 
 
 ## [0.2.0] - 2022-02-18